+++ R/data.R
@@ -75,7 +75,7 @@ NULL
 #' @docType data
 #'
 #' @description
-#' A `sf` object including all NUTS levels of Spain as provided by
+#' A \pkg{sf} object including all NUTS levels of Spain as provided by
 #' GISCO (2016 version).
 #'
 #' @source
@@ -88,17 +88,18 @@ NULL
 #' @format
 #' A `POLYGON` data frame (resolution: 1:1million, EPSG:4258) object with
 #' `r prettyNum(nrow(mapSpain::esp_nuts.sf), big.mark=",")` rows and fields:
-#'   * COAST_TYPE: COAST_TYPE
-#'   * FID: FID
-#'   * NUTS_NAME: NUTS name on local alphabet
-#'   * MOUNT_TYPE: MOUNT_TYPE
-#'   * NAME_LATN: Name on Latin characters
-#'   * CNTR_CODE: Eurostat Country code
-#'   * URBN_TYPE: URBN_TYPE
-#'   * NUTS_ID: NUTS identifier
-#'   * LEVL_CODE: NUTS level code (0,1,2,3)
-#'   * geometry: geometry field
-#'
+#' \describe{
+#'   \item{COAST_TYPE}{COAST_TYPE}
+#'   \item{FID}{FID}
+#'   \item{NUTS_NAME}{NUTS name on local alphabet}
+#'   \item{MOUNT_TYPE}{MOUNT_TYPE}
+#'   \item{NAME_LATN}{Name on Latin characters}
+#'   \item{CNTR_CODE}{Eurostat Country code}
+#'   \item{URBN_TYPE}{URBN_TYPE}
+#'   \item{NUTS_ID}{NUTS identifier}
+#'   \item{LEVL_CODE}{NUTS level code (0,1,2,3)}
+#'   \item{geometry}{geometry field}
+#' }
 #' @example inst/examples/esp_nuts_sf.R
 NULL
 
@@ -110,7 +111,7 @@ NULL
 #' @name esp_munic.sf
 #'
 #' @description
-#' A `sf` object including all municipalities of Spain as provided by GISCO
+#' A \pkg{sf} object including all municipalities of Spain as provided by GISCO
 #' (2019 version).
 #'
 #' @docType data
@@ -125,16 +126,17 @@ NULL
 #' @format
 #' A `POLYGON` data frame (resolution: 1:1million, EPSG:4258) object with
 #' `r prettyNum(nrow(mapSpain::esp_munic.sf), big.mark=",")` rows and fields:
-#'   * **codauto**: INE code of each autonomous community.
-#'   * **ine.ccaa.name**: INE name of each autonomous community.
-#'   * **cpro**: INE code of each province.
-#'   * **ine.prov.name**: INE name of each province.
-#'   * **cmun**: INE code of each municipality.
-#'   * **name**: Name of the municipality.
-#'   * **LAU_CODE**: LAU Code (GISCO) of the municipality. This is a
-#'     combination of **cpro** and **cmun**, aligned with INE coding scheme.
-#'   * **geometry**: geometry field.
-#'
+#' \describe{
+#'   \item{codauto}{INE code of each autonomous community.}
+#'   \item{ine.ccaa.name}{INE name of each autonomous community.}
+#'   \item{cpro}{INE code of each province.}
+#'   \item{ine.prov.name}{INE name of each province.}
+#'   \item{cmun}{INE code of each municipality.}
+#'   \item{name}{Name of the municipality.}
+#'   \item{LAU_CODE}{LAU Code (GISCO) of the municipality. This is a
+#'     combination of **cpro** and **cmun**, aligned with INE coding scheme.}
+#'   \item{geometry}{geometry field.}
+#' }
 #' @example inst/examples/esp_munic_sf.R
 NULL
 
@@ -160,48 +162,94 @@ NULL
 
 #' @title Public WMS and WMTS of Spain
 #'
-#' @family datasets
-#' @family imagery utilities
+#' @keywords internal
 #'
 #' @name leaflet.providersESP.df
 #'
 #' @description
+#' `r lifecycle::badge('superseded')`
+#'
+#' This data frame is not longer in use by \pkg{mapSpain}. See
+#' [esp_tiles_providers] instead.
+#'
 #' A data frame containing information of different public WMS and WMTS
 #' providers of Spain
 #'
-#' This function is a implementation of the javascript plugin
-#' [leaflet-providersESP](https://dieghernan.github.io/leaflet-providersESP/)
-#' **v1.3.0**.
 #'
 #' @docType data
 #'
 #' @source
 #' <https://dieghernan.github.io/leaflet-providersESP/> leaflet plugin,
-#' **v1.3.0**.
+#' **`r leafletprovidersESP_v`**.
 #'
 #' @encoding UTF-8
 #'
 #' @format
 #' A data frame object with a list of the required parameters for calling
 #' the service:
-#'   * **provider**: Provider name.
-#'   * **field**: Description of `value`.
-#'   * **value**: INE code of each province.
+#' \describe{
+#'   \item{provider}{Provider name}.
+#'   \item{field}{Description of `value`}.
+#'   \item{value}{INE code of each province}.
+#' }
 #'
+#' @examples
+#' data("leaflet.providersESP.df")
+NULL
+
+#' @title List with information of Public WMS and WMTS of Spain
+#'
+#' @family datasets
+#' @family imagery utilities
+#'
+#' @name esp_tiles_providers
+#'
+#' @description
+#' A named list of length `r length(esp_tiles_providers)` containing
+#' the parameters of the url information of different public WMS and WMTS
+#' providers of Spain.
+#'
+#' Implementation of javascript plugin
+#' [leaflet-providersESP](https://dieghernan.github.io/leaflet-providersESP/)
+#' **`r leafletprovidersESP_v`**.
+#'
+#' @docType data
+#'
+#' @source
+#' <https://dieghernan.github.io/leaflet-providersESP/> leaflet plugin,
+#' **`r leafletprovidersESP_v`**.
+#'
+#' @encoding UTF-8
+#'
+#' @format
+#' A named list of the providers available with the following structure:
+#' - Each item of the list is named with the provider alias.
+#' - Each element of the list contains two nested named lists:
+#'   - `static` with the parameters to get static tiles plus an additional item
+#'     named `attribution`.
+#'   - `leaflet` with additional parameters to be passed onto
+#'     [addProviderEspTiles()].
 #'
 #' @details
 #' Providers available to be passed to `type` on [esp_getTiles()] are:
 #'
-#' ```{r, echo=FALSE}
+#' ```{r, echo=FALSE, comment="", results="asis"}
 #'
-#' t <- mapSpain::leaflet.providersESP.df
-#' t <- paste0("'", unique(t$provider), "'")
-#' t <- data.frame(provider=t)
+#' t <- names(mapSpain::esp_tiles_providers)
+#' t <- paste0('\n - `"', t, '"`')
 #'
-#' knitr::kable(t)
+#'
+#' cat(t)
 #'
 #'
 #' ```
 #' @examples
-#' data("leaflet.providersESP.df")
+#' data("esp_tiles_providers")
+#' # Get a single provider
+#'
+#' single <- esp_tiles_providers[["IGNBase.Todo"]]
+#' single$static
+#'
+#' single$leaflet
+#'
 NULL