+++ R/esp_getTiles.R
@@ -6,7 +6,7 @@
 #'
 #' This function is a implementation of the javascript plugin
 #' [leaflet-providersESP](https://dieghernan.github.io/leaflet-providersESP/)
-#' **v1.3.0**.
+#' **`r leafletprovidersESP_v`**.
 #'
 #' @family imagery utilities
 #' @seealso [terra::rast()].
@@ -17,13 +17,17 @@
 #' .
 #' @source
 #' <https://dieghernan.github.io/leaflet-providersESP/> leaflet plugin,
-#'  **v1.3.0**.
+#'  **`r leafletprovidersESP_v`**.
 #'
 #' @export
 #'
-#' @param x An `sf` or `sfc` object.
+#' @param x An \pkg{sf} or `sfc` object.
 #'
-#' @param type Name of the provider. See [leaflet.providersESP.df].
+#' @param type This parameter could be either:
+#'   - The name of one of the  pre-defined providers
+#'     (see [esp_tiles_providers()]).
+#'   - A list with two named elements `id` and `q` with your own parameters.
+#'     See [esp_make_provider()] and examples.
 #' @param zoom Zoom level. If `NULL`, it is determined automatically. If set,
 #'   it overrides `zoommin`. Only valid for WMTS tiles. On a single point it
 #'   applies a buffer to the point and on `zoom = NULL` the function set a zoom
@@ -32,7 +36,7 @@
 #'   download fewer tiles than you probably want. Use `1` or `2` to
 #'   increase the resolution.
 #' @param crop `TRUE` if results should be cropped to the specified `x` extent,
-#'   `FALSE` otherwise. If `x` is an `sf` object with one `POINT`, crop is set
+#'   `FALSE` otherwise. If `x` is an \pkg{sf} object with one `POINT`, crop is set
 #'   to `FALSE`.
 #' @param res Resolution (in pixels) of the final tile. Only valid for WMS.
 #' @param bbox_expand A numeric value that indicates the expansion percentage
@@ -78,7 +82,7 @@
 #'
 #' ```
 #'
-#' For a complete list of providers see [leaflet.providersESP.df].
+#' For a complete list of providers see [esp_tiles_providers].
 #'
 #'
 #' Most WMS/WMTS providers provide tiles on "EPSG:3857". In case that the tile
@@ -86,20 +90,59 @@
 #'
 #' `x <- sf::st_transform(x, 3857)`
 #'
+#'
 #' @examples
 #' \dontrun{
 #' # This script downloads tiles to your local machine
 #' # Run only if you are online
 #'
-#' Murcia <- esp_get_ccaa_siane("Murcia", epsg = 3857)
-#' Tile <- esp_getTiles(Murcia)
+#' segovia <- esp_get_prov_siane("segovia", epsg = 3857)
+#' tile <- esp_getTiles(segovia)
 #'
 #' library(ggplot2)
 #' library(tidyterra)
 #'
-#' ggplot(Murcia) +
-#'   geom_spatraster_rgb(data = Tile) +
+#' ggplot(segovia) +
+#'   geom_spatraster_rgb(data = tile) +
 #'   geom_sf(fill = NA)
+#'
+#' # Another provider
+#'
+#' tile2 <- esp_getTiles(segovia, type = "MDT")
+#'
+#' ggplot(segovia) +
+#'   geom_spatraster_rgb(data = tile2) +
+#'   geom_sf(fill = NA)
+#'
+#' # A custom WMS provided
+#'
+#' custom_wms <- esp_make_provider(
+#'   id = "an_id_for_caching",
+#'   q = "https://idecyl.jcyl.es/geoserver/ge/wms?",
+#'   service = "WMS",
+#'   version = "1.3.0",
+#'   format = "image/png",
+#'   layers = "geolog_cyl_litologia"
+#' )
+#'
+#' custom_wms_tile <- esp_getTiles(segovia, custom_wms)
+#'
+#' autoplot(custom_wms_tile) +
+#'   geom_sf(data = segovia, fill = NA, color = "red")
+#'
+#' # A custom WMTS provider
+#'
+#' custom_wmts <- esp_make_provider(
+#'   id = "cyl_wmts",
+#'   q = "https://www.ign.es/wmts/pnoa-ma?",
+#'   service = "WMTS",
+#'   layer = "OI.OrthoimageCoverage"
+#' )
+#'
+#' custom_wmts_tile <- esp_getTiles(segovia, custom_wmts)
+#'
+#' autoplot(custom_wmts_tile) +
+#'   geom_sf(data = segovia, fill = NA, color = "white", linewidth = 2)
 #' }
 esp_getTiles <- function(x,
                          type = "IDErioja",
@@ -143,18 +186,6 @@ esp_getTiles <- function(x,
     )
   }
 
-  # A. Check providers
-  leafletProvidersESP <- mapSpain::leaflet.providersESP.df
-  provs <-
-    leafletProvidersESP[leafletProvidersESP$provider == type, ]
-
-  if (nrow(provs) == 0) {
-    stop(
-      "No match for type = '",
-      type,
-      "' found. Check providers available in mapSpain::leaflet.providersESP.df"
-    )
-  }
 
   # Some transformations
   res <- as.numeric(res)
@@ -163,13 +194,114 @@ esp_getTiles <- function(x,
   xinit <- x
   x <- sf::st_geometry(x)
 
-  # Transform to 3857 as it is native for tiles
 
-  x <- sf::st_transform(x, 3857)
+  # A. Check providers
+  if (is.list(type)) {
+    # Custom query
+
+    url_pieces <- type$q
+    type <- type$id
+
+    if (any(is.null(url_pieces), is.null(type))) {
+      stop("Custom provider should be a named list with an 'id' and a 'q' field")
+    }
+
+    url_pieces <- esp_hlp_split_url(url_pieces)
+    extra_opts <- NULL
+  } else {
+    provs <- mapSpain::esp_tiles_providers
+
+    if (!type %in% names(provs)) {
+      stop(
+        "No match for type = '",
+        type,
+        "' found. Check providers available in mapSpain::esp_tiles_providers"
+      )
+    }
+
+
+    # Split url
+    url_pieces <- provs[[type]]$static
+    # And get extra optios
+    extra_opts <- provs[[type]]$leaflet
+
+    names(url_pieces) <- tolower(names(url_pieces))
+    names(extra_opts) <- tolower(names(extra_opts))
+  }
+  # Create cache dir
+  cache_dir <- esp_hlp_cachedir(cache_dir)
+  cache_dir <- esp_hlp_cachedir(paste0(cache_dir, "/", type))
+
+
+  # Attribution
+  attr <- url_pieces$attribution
+
+  url_pieces <- modifyList(url_pieces, list(attribution = NULL))
+
+  # Get type of service
+  typeprov <- toupper(url_pieces$service)
+
+  # Add options
+  if (is.list(options)) {
+    names(options) <- tolower(names(options))
+
+    if (typeprov == "WMS" && "version" %in% names(options)) {
+      # Exception: need to change names depending on the version of WMS
+
+      v_wms <- unlist(modifyList(
+        list(v = url_pieces$version),
+        list(v = options$version)
+      ))
+
+
+      # Assess version
+      v_wms <- unlist(strsplit(v_wms, ".", fixed = TRUE))
+
+
+      if (v_wms[1] >= "1" && v_wms[2] >= "3") {
+        names(url_pieces) <- gsub("srs", "crs", names(url_pieces))
+      } else {
+        names(url_pieces) <- gsub("crs", "srs", names(url_pieces))
+      }
+    }
+
+    # Ignore TileMatrix fields in WMTS
+    if (typeprov == "WMTS") {
+      options <- options[!grepl("tilematrix", names(options), ignore.case = TRUE)]
+    }
+
+
+
+    url_pieces <- modifyList(url_pieces, options)
+    # Create new cache dir
+
+    # Modify cache dir
+    newdir <- paste0(names(options), "=", options, collapse = "&")
+    newdir <- esp_get_md5(newdir)
+
+    cache_dir <- file.path(cache_dir, newdir)
+    cache_dir <- esp_hlp_cachedir(cache_dir)
+  }
+
+
+
+  # Get CRS of Tile
+  crs <- unlist(url_pieces[names(url_pieces) %in% c("crs", "srs", "tilematrixset")])
+
+  if (tolower(crs) == tolower("GoogleMapsCompatible")) crs <- "epsg:3857"
+
+  crs_sf <- sf::st_crs(crs)
+
+
+  # Transform to crs of tile
+
+  x <- sf::st_transform(x, crs_sf)
 
   # Buffer if single point
   if (length(x) == 1 && "POINT" %in% sf::st_geometry_type(x)) {
-    x <- sf::st_buffer(sf::st_geometry(x), 50)
+    xmod <- sf::st_transform(sf::st_geometry(x), 3857)
+    xmod <- sf::st_buffer(xmod, 50)
+    x <- sf::st_transform(xmod, sf::st_crs(x))
     crop <- FALSE
     # Auto zoom = 15 if not set
     if (is.null(zoom)) {
@@ -178,52 +310,41 @@ esp_getTiles <- function(x,
     }
   }
 
-
-  # Create cache dir
-  cache_dir <- esp_hlp_cachedir(cache_dir)
-  cache_dir <- esp_hlp_cachedir(paste0(cache_dir, "/", type))
-
-  typeprov <- provs[provs$field == "type", "value"]
-
   newbbox <- esp_hlp_get_bbox(x, bbox_expand, typeprov)
-
 
   if (typeprov == "WMS") {
     rout <-
       getwms(
         newbbox,
-        provs,
+        url_pieces,
         update_cache,
         cache_dir,
         verbose,
         res,
-        transparent,
-        options
+        transparent
       )
   } else {
     rout <-
       getwmts(
         newbbox,
-        provs,
+        type,
+        url_pieces,
         update_cache,
         cache_dir,
         verbose,
-        res,
         zoom,
         zoommin,
-        type,
         transparent,
-        options
+        extra_opts
       )
   }
 
   # Regenerate
   # Display attributions
 
-  if (verbose) {
+  if (verbose && !is.null(attr)) {
     message(
-      "\nData and map tiles sources:\n",
-      provs[provs$field == "attribution_static", "value"]
+      "\nData and map tiles sources:\n", attr
     )
   }
 
@@ -232,7 +353,6 @@ esp_getTiles <- function(x,
 
   # reproject rout if needed
   if (!sf::st_crs(x) == sf::st_crs(rout)) {
-
     # Sometimes it gets an error
 
     rout_end <- try(terra::project(
@@ -271,7 +391,12 @@ esp_getTiles <- function(x,
   # Manage transparency
 
   if (!transparent && terra::nlyr(rout) == 4) {
-    rout <- rout[[-4]]
+    rout <- terra::subset(rout, 1:3)
+  }
+
+  # Manage RGB
+  if (isFALSE(terra::has.RGB(rout))) {
+    terra::RGB(rout) <- seq_len(terra::nlyr(rout))
   }
 
 
@@ -282,10 +407,6 @@ esp_getTiles <- function(x,
 #' Helper to get bboxes
 #' @noRd
 esp_hlp_get_bbox <- function(x, bbox_expand = 0.05, typeprov = "WMS") {
-  # Get bbox, this works with CRS 3857
-
-  stopifnot(identical(sf::st_crs(3857), sf::st_crs(x)))
-
   bbox <- as.double(sf::st_bbox(x))
   dimx <- (bbox[3] - bbox[1])
   dimy <- (bbox[4] - bbox[2])
@@ -314,4 +435,45 @@ esp_hlp_get_bbox <- function(x, bbox_expand = 0.05, typeprov = "WMS") {
   sf::st_crs(newbbox) <- sf::st_crs(x)
 
   return(newbbox)
+}
+
+# Helper to split urls
+esp_hlp_split_url <- function(url_static) {
+  split <- unlist(strsplit(url_static, "?", fixed = TRUE))
+
+  urlsplit <- list()
+  urlsplit$q <- paste0(split[1], "?")
+
+  opts <- unlist(strsplit(split[2], "&"))
+
+  names_opts <- vapply(opts, function(x) {
+    n <- strsplit(x, "=", fixed = TRUE)
+    return(unlist(n)[1])
+  }, FUN.VALUE = character(1))
+
+  values_opts <- vapply(opts, function(x) {
+    n <- strsplit(x, "=", fixed = TRUE)
+
+    unl <- unlist(n)
+    if (length(unl) == 2) {
+      return(unl[2])
+    }
+    return("")
+  }, FUN.VALUE = character(1))
+
+
+  names(values_opts) <- tolower(names_opts)
+
+  urlsplit <- modifyList(urlsplit, as.list(values_opts))
+
+  return(urlsplit)
+}
+
+esp_get_md5 <- function(x) {
+  tmp <- tempfile(fileext = ".txt")
+  writeLines(x, tmp)
+
+  md5 <- unname(tools::md5sum(tmp))
+
+  return(md5)
 }