+++ R/esp_get_railway.R
@@ -1,12 +1,12 @@
-#' Get `sf` lines and points with the railways of Spain
+#' Get \pkg{sf} lines and points with the railways of Spain
 #'
 #' @description
-#' Loads a `sf` lines or point object representing the nodes and railway
+#' Loads a \pkg{sf} lines or point object representing the nodes and railway
 #' lines of Spain.
 #'
 #' @family infrastructure
 #'
-#' @return A `sf` line or point object.
+#' @return A \pkg{sf} line or point object.
 #'
 #' @source IGN data via a custom CDN (see
 #' <https://github.com/rOpenSpain/mapSpain/tree/sianedata>).
@@ -44,7 +44,7 @@
 #'   geom_sf(data = ccaa, fill = NA) +
 #'   geom_sf(
 #'     data = rails, aes(color = tipo),
-#'     show.legend = FALSE, lwd = 1.5
+#'     show.legend = FALSE, linewidth = 1.5
 #'   ) +
 #'   geom_sf(
 #'     data = stations,