+++ tests/testthat/test-layer_spatraster.R
@@ -1,5 +1,4 @@
 test_that("Test layers", {
-
   # Test with sf
 
   x <- esp_get_ccaa("Catalonia")
@@ -14,8 +13,10 @@ test_that("Test layers", {
   skip_if_not_installed("terra")
   skip_if_not_installed("png")
 
+  # Skip test as tiles sometimes are not available
   skip_on_cran()
   skip_if_offline()
+
 
   tile <- esp_getTiles(x)